Fixed a B-tree edge case
When reading a size less than the size of the node, the math could return 0 blocks to be read instead of a minimum of 1.

Also fixed varying other little inaccuracies in the quest to hunt that one down and made some "critical" errors log as "error" so it would float up the stack properly.